+++ components/my-profile/create/index.ts
@@ -1,0 +1,17 @@
+import Header from "./header";
+import { Input } from "./input";
+import NameInput from "./name-input";
+import RadioGroup from "./radio-group";
+import GalleryButton from "./gallery-button";
+import DateTimeSheet from "./date-time-sheet";
+import BreedSelectSheet from "./breed-select-sheet";
+
+export {
+  Header,
+  Input,
+  NameInput,
+  RadioGroup,
+  GalleryButton,
+  DateTimeSheet,
+  BreedSelectSheet,
+}
